+++ primefaces/src/main/java/org/primefaces/component/treetable/TreeTableRenderer.java
@@ -586,6 +586,7 @@ public class TreeTableRenderer extends DataRenderer {
 
                 String columnStyleClass = column.getStyleClass();
                 String columnStyle = column.getStyle();
+                String title = column.getTitle();
                 int rowspan = column.getRowspan();
                 int colspan = column.getColspan();
                 int responsivePriority = column.getResponsivePriority();
@@ -617,6 +618,9 @@ public class TreeTableRenderer extends DataRenderer {
                 }
                 if (colspan != 1) {
                     writer.writeAttribute("colspan", colspan, null);
+                }
+                if (title != null) {
+                    writer.writeAttribute("title", title, null);
                 }
 
                 if (i == 0) {